configurações atualizadas
